requirements.txt
streamlit>=1.32
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0
faiss-cpu>=1.7
pytesseract>=0.3.10
pillow>=10.0.0
PyMuPDF>=1.23.0